+++ tests/base/value.test.js
@@ -64,6 +64,19 @@ describe("PeelrValue", function() {
     assert.equal(await val.extract("http://localhost:8000"), "h1 text content");
   });
 
+  it("calls onRequest when making requests", async function() {
+    let log = [];
+    let val = new PeelrValue("h1", {
+      onRequest: (params, cacheHit) => log.push([params, cacheHit])
+    });
+    val.getValue = function($el) {
+      return $el.text();
+    };
+
+    await val.extract("http://localhost:8000");
+    assert.deepEqual(log, [[{ url: "http://localhost:8000" }, false]]);
+  });
+
   it("extracts from request parameters", async function() {
     let val = new PeelrValue(".header#x-my-header");
     val.getValue = function($el) {
@@ -71,7 +84,7 @@ describe("PeelrValue", function() {
     };
     assert.equal(
       await val.extract({
-        uri: "http://localhost:8000/dump",
+        url: "http://localhost:8000/dump",
         headers: { "X-My-Header": "myvalue" }
       }),
       "myvalue"
@@ -111,6 +124,24 @@ describe("PeelrValue", function() {
       "item 4",
       "item 5",
       "item 6"
+    ]);
+  });
+
+  it("slices paginated values", async function() {
+    let val = new PeelrValue(".item", {
+      multiple: true,
+      nextPage: Peelr.attr(".next", "href"),
+      offset: 1,
+      limit: 4
+    });
+    val.getValue = function($el) {
+      return $el.text();
+    };
+    assert.deepEqual(await val.extract("http://localhost:8000/page1"), [
+      "item 2",
+      "item 3",
+      "item 4",
+      "item 5"
     ]);
   });
 });